Перемещение Abstraction в Application-слой
Все 6 файлов из слоя Core были перемещены в слой Application
