--- ChoreoCreatorBackend/ChoreoCreator.Core/Abstractions/IDancerPositionsRepository.cs
@@ -1,12 +1,0 @@
-﻿using ChoreoCreator.Core.Models;
-
-namespace ChoreoCreator.Core.Abstractions
-{
-    public interface IDancerPositionsRepository
-    {
-        Task<List<DancerPosition>> GetByFormationId(Guid formationId);
-        Task<Guid> Create(DancerPosition position);
-        Task<Guid> Update(Guid id, float x, float y);
-        Task<Guid> Delete(Guid id);
-    }
-}
