--- ChoreoCreatorBackend/ChoreoCreator.Core/Abstractions/IUsersService.cs
@@ -1,13 +1,0 @@
-﻿using ChoreoCreator.Core.Models;
-using ChoreoCreator.Core.ValueObjects;
-
-namespace ChoreoCreator.Core.Abstractions
-{
-    public interface IUsersService
-    {
-        Task<User?> GetByEmail(string email);  // Получить пользователя по email
-        Task<User?> ValidateCredentials(string email, string password);  // Проверка логина и пароля
-
-        Task<Result<UserId, string>> RegisterUser(string email, string password, CancellationToken ct);
-    }
-}
